Refactor tensor handling and enhance dtype consistency across core modules
- Updated the test checklist to reflect the completion status of integration tests, marking `test_end_to_end.py` as completed.
- Enhanced the `information_ricci.py` and `base_flow.py` files to ensure consistent handling of tensor dtypes, improving numerical stability and compatibility.
- Refactored the `ArithmeticDynamics` and `RiemannianFlow` classes to utilize complex weights and biases, ensuring proper initialization and dtype management.
- Improved the `NeuralQuantumBridge` class to handle both real and imaginary parts of amplitudes, enhancing quantum state representation.
- Streamlined tensor normalization processes in tests to ensure consistent dtype handling and improved validation of quantum state properties.

These changes enhance the reliability and organization of the codebase, aligning with best practices for tensor operations in quantum computing frameworks.

+++ src/core/patterns/arithmetic_dynamics.py
@@ -489,8 +489,8 @@ class ArithmeticDynamics(nn.Module):
         else:
             x_proj = x
 
-        # Convert to complex domain after interpolation
-        x_proj = x_proj.to(dtype=torch.complex64)
+        # Convert to the correct complex dtype
+        x_proj = x_proj.to(dtype=self.dtype)
 
         # Project to quantum state space
         quantum_state = self.quantum_proj(x_proj)  # [batch_size * ..., manifold_dim]
